requirements.txt
# Core Libraries
pyyaml
pandas
tqdm

# Computer Vision
opencv-python
easyocr

# Audio Analysis & Text
openai-whisper @ git+https://github.com/openai/whisper.git
librosa
thefuzz

# GUI Framework
PyQt6

# Enhanced functionality
Pillow
psutil